requirements.txt
requests>=2.31.0
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
joblib>=1.1.0
matplotlib>=3.4.0
statsmodels>=0.13.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
tenacity>=8.0.0
loguru>=0.6.0
backoff>=2.0.0
jsonschema>=4.0.0
pytest>=6.2.0
pytest-cov>=2.12.0
prefect>=2.0.0
great-expectations>=0.15.0
snowflake-connector-python>=3.0.0
cryptography>=40.0.0
